services/nlp-post/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
nats-py==2.6.0
pydantic==2.9.2
httpx==0.27.2
# transformers>=4.35.0  # Uncomment when using ByT5
# underthesea==6.8.1  # Vietnamese NLP toolkit
